Create population_adjustment
This code finds a factor to adjusts the population in a proportional way. The adjustment will be needed in the following cases:
a) when the allocated or forecasted population is less than the real population (census). 
b) when the allocated or forecasted population is higher than the real population (census).